chore: move mod_type to AgreementMod
