+++ backend/alembic/versions/2025_02_13_1646-8c9a30d8d6b4_model_changes_needed_to_load_data.py
@@ -1,8 +1,8 @@
-"""add tables needed for bli migration
+"""model changes needed to load data
 
-Revision ID: 8fbed03b0f3d
+Revision ID: 8c9a30d8d6b4
 Revises: 3d8237681112
-Create Date: 2025-02-12 22:11:08.848335+00:00
+Create Date: 2025-02-13 16:46:00.969438+00:00
 
 """
 from typing import Sequence, Union
@@ -10,9 +10,10 @@ from typing import Sequence, Union
 import sqlalchemy as sa
 from alembic import op
 from alembic_postgresql_enum import TableReference
+from sqlalchemy.dialects import postgresql
 
 # revision identifiers, used by Alembic.
-revision: str = '8fbed03b0f3d'
+revision: str = '8c9a30d8d6b4'
 down_revision: Union[str, None] = '3d8237681112'
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
@@ -23,6 +24,7 @@ def upgrade() -> None:
     op.create_table('agreement_mod_version',
     sa.Column('id', sa.Integer(), autoincrement=False, nullable=False),
     sa.Column('agreement_id', sa.Integer(), autoincrement=False, nullable=True),
+    sa.Column('mod_type', postgresql.ENUM('NEW', 'ADMIN', 'AMOUNT_TBD', 'AS_IS', 'REPLACEMENT_AMOUNT_FINAL', name='modtype', create_type=False), autoincrement=False, nullable=True),
     sa.Column('number', sa.String(), autoincrement=False, nullable=True),
     sa.Column('mod_date', sa.Date(), autoincrement=False, nullable=True),
     sa.Column('created_by', sa.Integer(), autoincrement=False, nullable=True),
@@ -105,6 +107,7 @@ def upgrade() -> None:
     op.create_table('agreement_mod',
     sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
     sa.Column('agreement_id', sa.Integer(), nullable=False),
+    sa.Column('mod_type', postgresql.ENUM('NEW', 'ADMIN', 'AMOUNT_TBD', 'AS_IS', 'REPLACEMENT_AMOUNT_FINAL', name='modtype', create_type=False), nullable=True),
     sa.Column('number', sa.String(), nullable=False),
     sa.Column('mod_date', sa.Date(), nullable=False),
     sa.Column('created_by', sa.Integer(), nullable=True),
@@ -159,10 +162,11 @@ def upgrade() -> None:
     op.add_column('budget_line_item', sa.Column('psc_fee_pymt_ref_nbr', sa.String(), nullable=True))
     op.add_column('budget_line_item', sa.Column('obligation_date', sa.Date(), nullable=True))
     op.create_foreign_key(None, 'budget_line_item', 'object_class_code', ['object_class_code_id'], ['id'])
-    op.create_foreign_key(None, 'budget_line_item', 'agreement_mod', ['mod_id'], ['id'])
     op.create_foreign_key(None, 'budget_line_item', 'ops_user', ['closed_by'], ['id'])
+    op.create_foreign_key(None, 'budget_line_item', 'agreement_mod', ['mod_id'], ['id'])
     op.drop_column('budget_line_item', 'requisition_date')
     op.drop_column('budget_line_item', 'requisition_number')
+    op.drop_column('budget_line_item', 'mod_type')
     op.add_column('budget_line_item_version', sa.Column('closed_by', sa.Integer(), autoincrement=False, nullable=True))
     op.add_column('budget_line_item_version', sa.Column('closed_date', sa.Date(), autoincrement=False, nullable=True))
     op.add_column('budget_line_item_version', sa.Column('extend_pop_to', sa.Date(), autoincrement=False, nullable=True))
@@ -176,21 +180,31 @@ def upgrade() -> None:
     op.add_column('budget_line_item_version', sa.Column('obligation_date', sa.Date(), autoincrement=False, nullable=True))
     op.drop_column('budget_line_item_version', 'requisition_date')
     op.drop_column('budget_line_item_version', 'requisition_number')
+    op.drop_column('budget_line_item_version', 'mod_type')
     op.drop_column('contract_agreement', 'invoice_line_nbr')
     op.drop_column('contract_agreement_version', 'invoice_line_nbr')
-    op.sync_enum_values('ops', 'modtype', ['NEW', 'ADMIN', 'AMOUNT_TBD', 'AS_IS', 'REPLACEMENT_AMOUNT_FINAL'],
-                        [TableReference(table_schema='ops', table_name='budget_line_item', column_name='mod_type'), TableReference(table_schema='ops', table_name='budget_line_item_version', column_name='mod_type')],
-                        enum_values_to_rename=[])
+    op.sync_enum_values(
+        enum_schema='ops',
+        enum_name='modtype',
+        new_values=['NEW', 'ADMIN', 'AMOUNT_TBD', 'AS_IS', 'REPLACEMENT_AMOUNT_FINAL'],
+        affected_columns=[TableReference(table_schema='ops', table_name='agreement_mod', column_name='mod_type'), TableReference(table_schema='ops', table_name='agreement_mod_version', column_name='mod_type')],
+        enum_values_to_rename=[],
+    )
     # ### end Alembic commands ###
 
 
 def downgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
-    op.sync_enum_values('ops', 'modtype', ['ADMIN', 'AMOUNT_TBD', 'AS_IS', 'REPLACEMENT_AMOUNT_FINAL'],
-                        [TableReference(table_schema='ops', table_name='budget_line_item', column_name='mod_type'), TableReference(table_schema='ops', table_name='budget_line_item_version', column_name='mod_type')],
-                        enum_values_to_rename=[])
+    op.sync_enum_values(
+        enum_schema='ops',
+        enum_name='modtype',
+        new_values=['ADMIN', 'AMOUNT_TBD', 'AS_IS', 'REPLACEMENT_AMOUNT_FINAL'],
+        affected_columns=[TableReference(table_schema='ops', table_name='agreement_mod', column_name='mod_type'), TableReference(table_schema='ops', table_name='agreement_mod_version', column_name='mod_type')],
+        enum_values_to_rename=[],
+    )
     op.add_column('contract_agreement_version', sa.Column('invoice_line_nbr', sa.INTEGER(), autoincrement=False, nullable=True))
     op.add_column('contract_agreement', sa.Column('invoice_line_nbr', sa.INTEGER(), autoincrement=False, nullable=True))
+    op.add_column('budget_line_item_version', sa.Column('mod_type', postgresql.ENUM('ADMIN', 'AMOUNT_TBD', 'AS_IS', 'REPLACEMENT_AMOUNT_FINAL', name='modtype', create_type=False), autoincrement=False, nullable=True))
     op.add_column('budget_line_item_version', sa.Column('requisition_number', sa.INTEGER(), autoincrement=False, nullable=True))
     op.add_column('budget_line_item_version', sa.Column('requisition_date', sa.DATE(), autoincrement=False, nullable=True))
     op.drop_column('budget_line_item_version', 'obligation_date')
@@ -204,6 +218,7 @@ def downgrade() -> None:
     op.drop_column('budget_line_item_version', 'extend_pop_to')
     op.drop_column('budget_line_item_version', 'closed_date')
     op.drop_column('budget_line_item_version', 'closed_by')
+    op.add_column('budget_line_item', sa.Column('mod_type', postgresql.ENUM('ADMIN', 'AMOUNT_TBD', 'AS_IS', 'REPLACEMENT_AMOUNT_FINAL', name='modtype', create_type=False), autoincrement=False, nullable=True))
     op.add_column('budget_line_item', sa.Column('requisition_number', sa.INTEGER(), autoincrement=False, nullable=True))
     op.add_column('budget_line_item', sa.Column('requisition_date', sa.DATE(), autoincrement=False, nullable=True))
     op.drop_constraint(None, 'budget_line_item', type_='foreignkey')
