+++ backend/models/agreements.py
@@ -32,6 +32,14 @@ class AgreementType(Enum):
     MISCELLANEOUS = auto()
 
 
+class ModType(Enum):
+    NEW = auto()
+    ADMIN = auto()
+    AMOUNT_TBD = auto()
+    AS_IS = auto()
+    REPLACEMENT_AMOUNT_FINAL = auto()
+
+
 class AgreementReason(Enum):
     NEW_REQ = auto()
     RECOMPETE = auto()  ## recompete is brand new contract related to same work
@@ -298,5 +306,6 @@ class AgreementMod(BaseModel):
     id: Mapped[int] = BaseModel.get_pk_column()
     agreement_id: Mapped[int] = mapped_column(ForeignKey("agreement.id"))
     agreement: Mapped[Agreement] = relationship("Agreement")
+    mod_type: Mapped[Optional[ModType]] = mapped_column(ENUM(ModType))
     number: Mapped[str] = mapped_column(String)
     mod_date: Mapped[date] = mapped_column(Date)
